Collapse message.HypervisorPortCheck into vmcheck.IsVirtualWorld
We still need to solve the PANIC issue.  If run in a non-virtual world,
the runtime will panic.
